margo rpc server bootstrap
This commit introduces a shared hostfile that is used
by servers to identify their rank when PMI2/PMIx is not
enabled. The unifycr command-line utility generates the
hostfile using the information collected from the resource
manager.

Each server processes the hostfile to initialize an
array of server information structs, indexed by server
rank. Currently, the info struct for each server contains
contains its hostname, MPI rank, and Margo address
string and hg_addr_t. The array is populated by querying
the key-value store for remote servers.

TEST_CHECKPATCH_SKIP_FILES=common/src/unifycr_configurator.h

+++ t/sharness.d/02-functions.sh
@@ -64,14 +64,30 @@ process_is_not_running()
 # Create metadata directory if needed and start daemon.
 unifycrd_start_daemon()
 {
+    # Make sure metadata directory exists
     if test -z "$UNIFYCR_META_DB_PATH"; then
+        return 1
+    elif ! test -d "$UNIFYCR_META_DB_PATH" &&
+         ! mkdir $UNIFYCR_META_DB_PATH; then
         return 1
     fi
 
-    if ! test -d "$UNIFYCR_META_DB_PATH" &&
-       ! mkdir $UNIFYCR_META_DB_PATH; then
-            return 1
-    fi
+    # Generate servers hostfile
+    # if test -z "$UNIFYCR_SHAREDFS_DIR"; then
+    #     return 1
+    # elif ! test -d "$UNIFYCR_SHAREDFS_DIR" &&
+    #      ! mkdir $UNIFYCR_SHAREDFS_DIR; then
+    #     return 1
+    # fi
+    # srvr_hosts=$UNIFYCR_SHAREDFS_DIR/unifycrd.hosts
+    # if [ ! -f $srvr_hosts ]; then
+    #     touch $srvr_hosts
+    #     echo "1" >> $srvr_hosts
+    #     hostname >> $srvr_hosts
+    # fi
+    # export UNIFYCR_SERVER_HOSTFILE=$srvr_hosts
+
+    # run server daemon
     $UNIFYCRD
 }
 
@@ -84,5 +100,6 @@ unifycrd_stop_daemon()
 # Remove the metadata directory.
 unifycrd_cleanup()
 {
-    rm -rf $UNIFYCR_META_DB_PATH
+    test -d "$UNIFYCR_META_DB_PATH" && rm -rf $UNIFYCR_META_DB_PATH
+    # test -d "$UNIFYCR_SHAREDFS_DIR" && rm -rf $UNIFYCR_SHAREDFS_DIR
 }
